Move range string tests to test_rangestring
